ui fixes

+++ fwptt.TestProject/Project/TimeLine/TimeLineStatus.cs
@@ -8,12 +8,12 @@ namespace fwptt.TestProject.Project.TimeLine
 {
     public class TimeLineStatus
     {
-        public int MaxExecutionThreads { get; set; }
+        public uint MaxExecutionThreads { get; set; }
         public int MiliSecondsPauseBetweenRequests { get; set; }
-        public int CurrentExecutionThreads { get; set; }
+        public uint CurrentExecutionThreads { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
         public virtual bool IsRunning { get; set; }
-        public int CurrentIteration { get; set; }
+        public ulong CurrentIteration { get; set; }
     }
 }
